feat: abstract signup form
